[UI]:공용 모달 컴포넌트 구현
